4. Archive/25S2 ; 단국대학교 4학년 2학기/인공지능융합플랫폼/포트폴리오 서비스 구현/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0

# Google Gemini 공식 SDK
google-generativeai>=1.0.0

# JSON5 파싱
json5>=0.9.0

# HTML 파싱 (필요 시)
beautifulsoup4>=4.12.0

# PDF 생성을 위한 Playwright
playwright>=1.45.0